Added new tests and fixed evaluator
Added new tests for when the handler is used in the step, segment and map functions.

Changed some test names to be more descriptive.

When evaluator's free variable count is 1 the handler receives a scalar as the positions (first argument) otherwise it will receive a vector.

When the evaluator's dimension is 1 the handler receives a scalar as the value argument (second argument) otherwise a vector.